Finish refactor and pass test with working app.
