+++ actions/sbotGatheringWasAdded.js
@@ -5,7 +5,7 @@ import Model from '../models/model'
 const SbotGatheringAdded = t.struct({payload: Gathering}, 'sbotGatheringWasAdded')
 SbotGatheringAdded.prototype.update = function(model) {
 	return {model: Model.update(model, 
-		{events: {$push: [this.payload]}}	
+		{gatherings: {$push: [this.payload]}}	
 	)}
 }
