+++ test/actions/sbotGatheringWasAdded.js
@@ -6,13 +6,13 @@ import Gathering from '../../models/gathering'
 import Model from '../../models/model'
 
 
-test('push new event into events', function(t) {
+test('push new gathering into gatherings', function(t) {
   const model = Model({url: '', rsvps: [], gatherings: Gatherings([validGathering])}) 
   const newGathering = Gathering(validGathering)
   const action = sbotGatheringWasAdded({payload: newGathering})
 
   const newModel = action.update(model)
-  t.equal(newModel.model.gatherings.length, 2, 'new events has length 2')
+  t.equal(newModel.model.gatherings.length, 2, 'new gatherings has length 2')
   t.end()
 })
