+++ test/actions/sbotMyRsvpWasAdded.js
@@ -6,7 +6,7 @@ import Model from '../../models/model'
 
 
 test('if rsvp link is new, push new rsvp into rsvps', function(t) {
-  const model = Model({url: '', events: [], rsvps: Rsvps([{link: 'dkjfd', value: 1}])}) 
+  const model = Model({url: '', gatherings: [], rsvps: Rsvps([{link: 'dkjfd', value: 1}])}) 
   const newRsvp = Rsvp({link:'piet', value: 0})
   const action = sbotMyRsvpWasAddedAction({payload: newRsvp})
 
@@ -16,7 +16,7 @@ test('if rsvp link is new, push new rsvp into rsvps', function(t) {
 })
 
 test('if rsvp link already exists, replace old rsvp', function(t) {
-  const model = Model({url: '', events: [], rsvps: Rsvps([{link: 'dkjfd', value: 1}])}) 
+  const model = Model({url: '', gatherings: [], rsvps: Rsvps([{link: 'dkjfd', value: 1}])}) 
   const newRsvp = Rsvp({link:'dkjfd', value: 0})
   const action = sbotMyRsvpWasAddedAction({payload: newRsvp})
